Simplify Canvas landing page
