+++ apps/canvas/src/features/shared/components/top-navigation.tsx
@@ -1,6 +1,4 @@
-import React from "react";
 import { Link } from "react-router-dom";
-import { Bell } from "lucide-react";
 import { Button } from "@/design-system/ui/button";
 import { cn } from "@/lib/utils";
 import ActiveWorkspaceIndicator from "@/features/shared/components/top-navigation/active-workspace-indicator";
@@ -34,9 +32,6 @@ export default function TopNavigation({
         <VersionStatus />
         <Button variant="outline" size="sm" asChild>
           <Link to="/workflow-remediations">Remediations</Link>
-        </Button>
-        <Button variant="ghost" size="icon">
-          <Bell className="h-5 w-5" />
         </Button>
         <AccountMenu
           credentials={credentials}
