+++ apps/canvas/src/features/workflow/pages/workflow-gallery/types.ts
@@ -1,23 +1,3 @@
 export type WorkflowGalleryTab = "all" | "pinned" | "templates";
 
-export type WorkflowGallerySort = "updated" | "created" | "name";
-
 export type WorkflowGalleryTabCounts = Record<WorkflowGalleryTab, number>;
-
-export interface WorkflowGalleryFilters {
-  owner: {
-    me: boolean;
-    shared: boolean;
-  };
-  status: {
-    active: boolean;
-    draft: boolean;
-    archived: boolean;
-  };
-  tags: {
-    favorite: boolean;
-    template: boolean;
-    production: boolean;
-    development: boolean;
-  };
-}
