+++ apps/canvas/src/features/workflow/pages/workflow-gallery/use-workflow-gallery.ts
@@ -5,7 +5,6 @@ export const useWorkflowGallery = () => {
   const state = useWorkflowGalleryState();
   const actions = useWorkflowGalleryActions({
     setSelectedTab: state.setSelectedTab,
-    setShowFilterPopover: state.setShowFilterPopover,
   });
 
   return {
